Factor get-source-url and add comments
